Add a controller to deal with any OPTIONS requests
The OPTIONS request is used by CORS (Cross-Origin
Resource Sharing) for so-called "preflight" checks.
I'm coming to the conclusion that this not a useful
commit - it mostly checks that the origin is the
same as request.host, in which case this shouldn't
be a cross-origin request anyway.  Google translate
seems to work fine even with the requests being
refused.

Ditch this commit onto a topic branch in case it
ever becomes useful.